# Stop using unic crate and support Unicode 13 (fixes #497)