--- a/Makefile
+++ b/Makefile
@@ -99,4 +99,7 @@ update-stencil:
 	$(PYTHON) update_stencil.py \
 		../mozilla-unified ./
 
+update-unicode:
+	$(PYTHON) update_unicode.py UNIDATA ./
+
 .PHONY: all check static-check dyn-check jsdemo rust update-opcodes-m-u