--- a/crates/parser/Cargo.toml
+++ b/crates/parser/Cargo.toml
@@ -10,7 +10,6 @@ bumpalo = "2.6.0"
 jsparagus-ast = { path = "../ast" }
 jsparagus-json-log = { path = "../json-log" }
 jsparagus-generated-parser = { path = "../generated_parser" }
-unic-ucd-ident = { version = "0.9.0", features = ["id"] }
 
 [features]
 logging = ["jsparagus-json-log/logging"]