… Cargo.lock
… crates/parser/src/unicode_data.rs