Instead of using `is_id_start`/`is_id_continue` in `unic_ucd_ident`,
generate the table from `DerivedCoreProperties.txt`, in the same way as m-c does
https://searchfox.org/mozilla-central/source/js/src/util/make_unicode.py

Pros:
  * Easy to bump Unicode version
  * Better performance in most case (see below)

Cons:
  * Maintenance cost
  * Code duplication between jsparagus and mozilla-central (some python functions are copied from m-c)

This improves the performance for BMP, ~~especially for ASCII range~~.
while regresses on non-BMP (it's expected. no optimization is applied to non-BMP, given we don't expect non-BMP frequently appears inside code)

|        | range   | time (ns per call) | actual result |
|--------|---------|---------------|-----------------------------------------------|
| before | ASCII   | 74 | (0.95s for 100000 loops over U+0000 - U+007F) |
| before | BMP     | 70 | (45.85s for 10000 loops over U+0000 - U+FFFF) |
| before | non-BMP | 69 | (6.33s for 100 loops over U+10000 - U+F0000)  |
|after |ASCII | 1.6 | (0.02s for 100000 loops over U+0000 - U+007F) |
|after |BMP | 3.1 | (2.04s for 10000 loops over U+0000-U+FFFF) |
|after |non-BMP | 132| (12.12s for 100 loops over U+10000 - U+F0000) |