@arai-a 2020-06-03
> especially for ASCII range.

Scratch that.
we aren't using is_id_start/is_id_continue in unic_ucd_ident for ASCII range.