@arai-a 2020-06-03
to be clear, the above score is for `is_id_start` and ` is_id_continue` itself, before/after the patch.

we're not using `id_id_start` and `is_id_continue` for ASCII range for perf reason,
https://github.com/mozilla-spidermonkey/jsparagus/blob/573a588f04695e9e0a593713f2196ce59aaa9374/crates/parser/src/lexer.rs#L272-L279
https://github.com/mozilla-spidermonkey/jsparagus/blob/573a588f04695e9e0a593713f2196ce59aaa9374/crates/parser/src/lexer.rs#L294-L301

after the change, we can remove the workaround in `is_identifier_start` etc and directly call `is_id_start`.